feat: parallel batch [T20260509-4]
Tasks:
- T20260509-4: Modularize crates/orbit-core/src/runtime/orbit_tool_host/mod.rs (2,033 lines)
